Fix of FS#1
In case of multiple inclusion levels in an inclusion hierarchy, the traversal
of it let to the content of the last visited leaf being prepended to the rendition
of the parent node thanks to a variable initialized at the wrong location in code.
Relocating the initialization of $this->doc to AFTER traversal fixes this problem.

darcs-hash:20080408141449-2b4f5-05724ca31f6ca842c143636c3178920783235729.gz