New attempt of a rewrite using instructions during rendering
This patch makes the include plugin no longer do all the work in the
instructions but instead during rendering the instructions of the child
pages are converted and executed in the renderer of the page itself.
This means all advantages of working at instruction level are preserved,
but tricky caching problems are avoided. This is still not working
completely:

- The cache doesn't work at all, if your metadata is fresh the renderer
  cache won't be used at all
- No tests have been done if ACL support is actually working (but it
  should)
- Problems with the box plugin that have been fixed might reappear
- The support for different levels hasn't been tested that thoroughly
- The blog plugin will be broken with this version, perhaps
  compatibility fixes will be added again later
- The metadata cache isn't refreshed when it should, at least in the
  current git version of DokuWiki