Add section edit button handling
This commit fixes section edit buttons. Now section edit buttons of
included pages don't interfere with buttons of the parent page anymore
and all section edits of the included page are closed at the end of the
included page. Furthermore the id and revision of section edit buttons
of included pages are correct so they should actually work. A
redirect_id like in other edit buttons ensures that the user is
redirected to the correct page after the edit.