Wrap content that is not in a section in a section
At the beginning of every page there might be content that is not in a
section, that is the content before the first header. When such a page
is included in another page at a higher level this content without a
section hasn't been indented properly. Now when needed a new section
around that content is created.