Force metadata rendering in older DokuWiki versions
With this change the include plugins forces metadata rendering when it
redirects the user after saving a page to another page because otherwise
the metadata is only rendered when the page is displayed directly in
DokuWiki versions older than 2010-11-23. A version check ensures this
code is only executed in older DokuWiki versions.