Add inline mode without sections for inclusion in other plugins
This mode might look strange with some templates as it changes e.g. how
headers are displayed. If the included content contains no headers this
mode shouldn't cause any visual difference. If you have problems with
invalid html or the inclusion of included content in wrap-blocks of the
wrap-plugin or in hidable blocks try the inline mode, it should solve
these problems.