Add linkonly option for displaying a link instead of the content
This commit adds a linkonly option that makes the include plugin behave
like internal links except that each link is wrapped inside a paragraph.
When the url parameter "include_content" is set, all includes are
displayed as full includes. This makes it possible to have a page with a
list of a lot of pages as table of contents and then to export all these
pages as single large document with setting that url parameter. This is
useful for creating things like books or larger documentations.
This feature has been requested and sponsored by Jörg Ferlein.