Use sectionID() only when a section is given. Closes #67 and closes #66
The problem was that sectionID() when NULL is given as parameter returns
'section' which leads to strange links in headings and empty headings
also get 'section' as id so they are selected when the whole page is
included. Please upgrade and purge the cache (at least for the pages
where the bug occurs).