Fix loading instructions with the wrong global $ID
Previously, instructions have been loaded while $ID has been set to the
current page instead of the page that has been loaded. Now this is the
page that is loaded as some plugins like the discussion plugin do save
data for the wrong page otherwise. This should fix some problems in the
discussion plugin where discussion sections appeared on the wrong page.