Add more options for the linkonly mode and date replacements
This adds several new options and flags, it allows to force-enable the
usage of the first heading of the linked to page even if useheading is
off, it allows to disable the paragraph around the links and it adds
more date replacements so you can easily use the year of the previous or
next month etc. in the id of the included page/namespace.