Make the edit page button use section edits, fixes #65
This means that now the edit buttons for included pages only appear
where it makes sense (i.e. not in the preview and not in the RSS feed).
This also means that the cache no longer depends on the edit permissions
as there is nothing in the rendering process anymore that depends on
them because the edit buttons are generated after the rendering.