Set the end position of the last section edit button of section includes
When not the whole page but just a section was included the last section
edit button inside the included page was for the whole rest of the page
instead of the rest of the included section. Now the position after the
included content is stored and used in the section edit button.