Add "safeindex" feature, prevents indexing of protected included metadata
The safeindex feature that is turned on by default prevents the indexer
from indexing metadata from included pages that are non-public. This
means that for example only links from included pages that are public
will be indexed.

This affects plugins that add their own metadata to the index in the
following ways:

* there is no effect when all included pages are public or when no user
  is logged in when the page is indexed
* when the plugin's event handler is called after the include plugin,
  the plugin will get only metadata from included pages that are public
* when the plugin's event handler is called before the include plugin
  the include plugin will delete the plugin's metadata. I'm happy to add
  special handlers or exceptions for plugins like the tag plugin
  (already included) that are affected by this problem.

The safeindex feature can be turned off when the ACL rules of all parent
pages match the child pages or when information disclosure through
metadata like backlinks is no problem.