Fix #137 including wrap sections with header
When the included page is indented the include plugin opens a new
section in it if the page doesn't begin with a header. This section was
closed directly before the next header as the include plugin assumed
that a new section would be opened after the header. However when this
header is inside a wrap section of the wrap plugin there is no section
open after the header and the section close destroys the HTML structure.
This change makes sure that the section is only closed if there is also
a new section opened.