Version 2018-11-29: Security fix for DokuWiki Greebo
If you are using the DokuWiki Greebo release and rely on ACL checks in
the include plugin, apply this change as soon as possible. Note that
this is only an issue with namespace includes, so if you do not use
namespace includes and edits are only allowed for users that have access
to your whole wiki, this does not concern you (but updating is still
recommended).

Note that this is a problem caused by a bug in DokuWiki release Greebo.
A future hotfix release of DokuWiki might fix this, too, see
https://github.com/splitbrain/dokuwiki/pull/2609 for further
information.