New noinstructionscache flag
This commit adds a new flag: noinstructionscache to plugin syntax. This flag forces not to use an instruction cache for the included page. This flag allows using the struct serial aggregation inside the included page.
More generally, it allows using any DokuWiki syntax which instructions are dependant on the current page id, inside the included page.