clean up helper
(auto) reformatting
PSR-2 renaming

I started on cleaning up the instruction handling methods, but left most
of them as is. This needs a more detail oriented refactoring. The
passing of an instruction array that is modified by reference is not
ideal.

Methods still need type hints and parameter doc blocks.